Radi mi ažuriranje, popravio vrste podataka za neke atribute, imam mali problemčić kod naziva ali to će sve lako biti riješeno u idućoj verziji

+++ Software/EkoInventar/Models/Materijal.cs
@@ -11,9 +11,9 @@ namespace EkoInventar.Models
 
         public int ID_mat { get; set; }
         public string Naziv {  get; set; }
-        public int Cijena { get; set; }
+        public float Cijena { get; set; }
         public int Kolicina { get; set; }
-        public int KriticniPostotak { get; set; }
+        public float KriticniPostotak { get; set; }
         public int maxKolicina { get; set; }
     }
 }
